PDF/Demo1/PDF/requirements.txt
Flask==3.1.2
gunicorn==22.0.0
pdf2image==1.17.0
pandas==2.3.3
openpyxl==3.1.5
pytesseract==0.3.13
pillow==11.3.0
numpy==2.2.6
lxml==6.0.2
pdfminer.six==20250506
flask_cors==4.0.1
requests==2.32.5
